+++ restflow/responses/__init__.py
@@ -1,3 +1,4 @@
+from restflow.responses.response import Response
 from restflow.responses.streaming import (
     NDJSONResponse,
     SSEResponse,
@@ -6,6 +7,7 @@ from restflow.responses.streaming import (
 
 __all__ = [
     "NDJSONResponse",
+    "Response",
     "SSEResponse",
     "StreamingJSONListResponse",
 ]